\name{oaks}
\alias{oaks}
\docType{data}
\title{
Dataset of oak observations and names of characters collected
}
\description{
These are measurements on individual leaves collected from across the range of
bur oak (_Quercus macrocarpa_).
}
\usage{data("oaks")}
\format{
A list of 2 elements
 $ charVec: a vector of characters in the traits data frame
 $ traits : a data frame of measurements made on 272 leaves from several trees at each of numerous sites
}
\details{
Measurements were made on bur oak (_Quercus macrocarpa_). Details in study cited under references.
}
\source{
To be added after publication
}
\references{
To be added after publication
}
\examples{
data(oaks)
}
\keyword{datasets}
