\name{oaks}
\alias{oaks}
\docType{data}
\title{
Dataset of oak observations and names of characters collected
}
\description{
These are measurements on individual leaves collected from across the range of
bur oak (_Quercus macrocarpa_).
}
\usage{data("oaks")}
\format{
A list of 2 elements
 $ charVec: a vector of characters in the traits data frame
 $ traits : a data frame of measurements made on 272 leaves from several trees at each of numerous sites
}
\details{
Measurements were made on bur oak (_Quercus macrocarpa_). Details in study cited under references.
}
\source{
Desmond SC, Garner M, Flannery S, Whittemore AT, Hipp AL. 2020. Leaf shape and size variation in bur oaks: An empirical study and simulation of sampling strategies. bioRxiv: 2020.05.11.088039.
}
\references{
Desmond SC, Garner M, Flannery S, Whittemore AT, Hipp AL. 2020. Leaf shape and size variation in bur oaks: An empirical study and simulation of sampling strategies. bioRxiv: 2020.05.11.088039.
}
\examples{
data(oaks)
}
\keyword{datasets}
