\name{traitsPopSim-package}
\alias{traitsPopSim-package}
\alias{traitsPopSim}
\docType{package}
\title{
\packageTitle{traitsPopSim}
}
\description{
\packageDescription{traitsPopSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{traitsPopSim}
\packageIndices{traitsPopSim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{traitsPopSim}

Maintainer: \packageMaintainer{traitsPopSim}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
