\name{traitsPopSim-package}
\alias{traitsPopSim-package}
\alias{traitsPopSim}
\docType{package}
\title{
\packageTitle{traitsPopSim}
}
\description{
\packageDescription{traitsPopSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{traitsPopSim}
\packageIndices{traitsPopSim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{traitsPopSim}

Maintainer: \packageMaintainer{traitsPopSim}
}
